Refactored graph classes to avoid code duplication in the future; Added building deep slivers from graph model and associated test code
